Cookies serialization moved to SimpleTableSerializer
