+++ src/main/scala/org/fs/mael/backend/http/ui/HeadersFieldEditorDialog.scala
@@ -3,7 +3,7 @@ package org.fs.mael.backend.http.ui
 import scala.collection.immutable.ListMap
 
 import org.eclipse.swt.widgets._
-import org.fs.mael.backend.http.HttpUtils
+import org.fs.mael.backend.http.utils.HttpUtils
 import org.fs.mael.core.utils.CoreUtils._
 import org.fs.mael.ui.components.StringTablePopupEditorDialog
 import org.slf4s.Logging
